+++ src/pages/task/finishTask/DetailTaskPage.tsx/CancelTask.tsx
@@ -8,10 +8,7 @@ import color from "../../../../resource/color";
 import GoogleMap from "../../../../components/map/GoogleMap";
 import { LAT_LNG_BANGKOK } from "../../../../definitions/Location";
 import TextArea from "antd/lib/input/TextArea";
-import {
-  CalendarOutlined,
-  ClockCircleOutlined,
-} from "@ant-design/icons";
+import { CalendarOutlined, ClockCircleOutlined } from "@ant-design/icons";
 import {
   DetailFinishTask,
   DetailFinishTask_INIT,
@@ -24,7 +21,10 @@ import {
   HistoryEntity_INIT,
 } from "../../../../entities/HistoryEntities";
 import { CouponDataSource } from "../../../../datasource/CouponDatasource";
-import { numberWithCommas } from "../../../../utilities/TextFormatter";
+import {
+  numberWithCommas,
+  numberWithCommasToFixed,
+} from "../../../../utilities/TextFormatter";
 import {
   STATUS_EN_NEWTASK_COLOR_MAPPING,
   STATUS_NEWTASK_MAPPING,
@@ -57,14 +57,15 @@ function CancelTask() {
         CouponDataSource.getPromotionCode(res.data.couponId).then((result) =>
           setCouponData({
             couponCode: res.data.couponCode ?? "",
-            couponDiscount: !res.data.discount ? null : parseInt(res.data.discount),
+            couponDiscount: !res.data.discountCoupon
+              ? null
+              : parseInt(res.data.discountCoupon),
             couponName: result.couponName ?? "",
           })
         );
       }
       setHistory(res.data.taskHistory[0]);
       setData(res);
-      console.log("res", res);
       setMapPosition({
         lat: parseFloat(res.data.farmerPlot.lat),
         lng: parseFloat(res.data.farmerPlot.long),
@@ -431,19 +432,24 @@ function CancelTask() {
     <Form style={{ padding: "20px" }}>
       <Form style={{ padding: "20px", backgroundColor: "#2196531A" }}>
         <div className="row">
-          <div className="col-lg">
-            <Form.Item>
-              <span>
-                ยอดรวมค่าบริการ
-                <br />
-                <b style={{ fontSize: "20px", color: color.Success }}>
-                  {data.data.totalPrice !== null
-                    ? numberWithCommas(parseFloat(data.data.totalPrice)) +
-                      " บาท"
-                    : "0 บาท"}
-                </b>
-              </span>
-            </Form.Item>
+          <div className="col-lg-3" style={{ borderRight: "solid" }}>
+            <label>ยอดรวมค่าบริการ (เกษตรกร)</label>
+            <h5 style={{ color: color.primary1 }} className="p-2">
+              {data?.data.totalPrice &&
+                numberWithCommasToFixed(parseFloat(data?.data.totalPrice))}{" "}
+              บาท
+            </h5>
+          </div>
+          <div className="col-lg-3" style={{ paddingLeft: "40px" }}>
+            <label>รายได้ที่นักบินโดรนได้รับ</label>
+            <h5 style={{ color: color.Warning }} className="p-2">
+              {data?.data.price &&
+                numberWithCommasToFixed(
+                  parseFloat(data?.data.price) +
+                    parseFloat(data?.data.revenuePromotion)
+                )}{" "}
+              บาท
+            </h5>
           </div>
         </div>
         <div className="row">
@@ -489,6 +495,8 @@ function CancelTask() {
               />
             </Form.Item>
           </div>
+        </div>
+        <div className="row">
           <div className="form-group col-lg-4">
             <label>รหัสคูปอง</label>
             <Input value={couponData.couponCode} disabled autoComplete="off" />
@@ -500,7 +508,28 @@ function CancelTask() {
           <div className="form-group col-lg-4">
             <label>ส่วนลดคูปอง</label>
             <Input
+              suffix="บาท"
               value={numberWithCommas(couponData.couponDiscount!)}
+              disabled
+              autoComplete="off"
+            />
+          </div>
+        </div>
+        <div className="row pt-3">
+          <div className="form-group col-lg-6 p-2">
+            <label>โปรโมชั่นนักบินโดรน</label>
+            <Input
+              suffix="บาท"
+              value={data.data.discountPromotion || 0}
+              disabled
+              autoComplete="off"
+            />
+          </div>
+          <div className="form-group col-lg-6 p-2">
+            <label>โปรโมชั่นเกษตรกร</label>
+            <Input
+              suffix="บาท"
+              value={data.data.revenuePromotion || 0}
               disabled
               autoComplete="off"
             />